mail.operas.pt/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==3.0.10